+++ models/conformer/encoder_layer.py
@@ -12,11 +12,11 @@ import torch
 
 from torch import nn, Tensor
 
-from modules.conformer.convolution import ConvolutionModule
-from modules.transformer.attention import MultiHeadedAttention, RelPositionMultiHeadedAttention
-from modules.transformer.layer_norm import LayerNorm
-from modules.transformer.multi_layer_conv import Conv1dLinear, MultiLayeredConv1d
-from modules.transformer.positionwise_feed_forward import PositionwiseFeedForward
+from models.conformer.convolution import ConvolutionModule
+from models.transformer.attention import MultiHeadedAttention, RelPositionMultiHeadedAttention
+from models.transformer.layer_norm import LayerNorm
+from models.transformer.multi_layer_conv import Conv1dLinear, MultiLayeredConv1d
+from models.transformer.positionwise_feed_forward import PositionwiseFeedForward
 
 
 class EncoderLayer(nn.Module):
