+++ models/variance_predictor.py
@@ -3,7 +3,7 @@ from torch import nn, Tensor
 from typing import TypedDict, Optional
 from torchtyping import TensorType
 
-from modules.transformer.layer_norm import LayerNorm
+from models.transformer.layer_norm import LayerNorm
 
 
 class VariancePredictorConfig(TypedDict):
